Fix signed vs. unsigned comparison warning

+++ freecusd/lib.c
@@ -556,7 +556,7 @@ int fcd_lib_disk_index(char c)
 {
 	int i;
 
-	for (i = 0; i < fcd_conf_disk_count; ++i) {
+	for (i = 0; i < (int)fcd_conf_disk_count; ++i) {
 
 		if (c == fcd_conf_disk_names[i][FCD_DISK_NAME_SIZE - 2])
 			return i;
